mqqt-app/requirements.txt
paho-mqtt==1.6.1
SQLAlchemy==2.0.21 
pymysql==1.1.0
